Ajuste componente alertas - loop infinito AB#9374 AB#4513
