+++ src/SME.SGP.WebClient/src/componentes-sgp/conteudo.js
@@ -1,17 +1,17 @@
+import { Modal } from 'antd';
 import React, { useEffect, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { Switch } from 'react-router-dom';
 import shortid from 'shortid';
-import { useSelector, useDispatch } from 'react-redux';
-import { Modal, Row } from 'antd';
 import styled from 'styled-components';
-import { alertaFechar } from '../redux/modulos/alertas/actions';
+import rotasArray from '~/rotas/rotas';
+
 import Button from '../componentes/button';
 import { Colors } from '../componentes/colors';
-import BreadcrumbSgp from './breadcrumb-sgp';
-import Alert from '~/componentes/alert';
-import Grid from '~/componentes/grid';
-import rotasArray from '~/rotas/rotas';
+import { alertaFechar } from '../redux/modulos/alertas/actions';
 import RotaAutenticadaEstruturada from '../rotas/rotaAutenticadaEstruturada';
+import BreadcrumbSgp from './breadcrumb-sgp';
+import Mensagens from './mensagens/mensagens';
 
 const ContainerModal = styled.div`
   .ant-modal-footer {
@@ -39,8 +39,6 @@ const Conteudo = () => {
     confirmacao.resolve(resultado);
     dispatch(alertaFechar());
   };
-
-  const alertas = useSelector(state => state.alertas);
 
   return (
     <div style={{ marginLeft: retraido ? '115px' : '250px' }}>
@@ -82,31 +80,7 @@ const Conteudo = () => {
               <b>{confirmacao.textoNegrito}</b>
             </Modal>
           </ContainerModal>
-          <div className="card-body m-r-0 m-l-0 p-l-0 p-r-0 m-t-0">
-            {alertas.alertas.map(alerta => (
-              <Row key={shortid.generate()}>
-                <Grid cols={12}>
-                  <Alert alerta={alerta} key={alerta.id} closable />
-                </Grid>
-              </Row>
-            ))}
-            <Row
-              key={shortid.generate()}
-              hidden={!menuRetraido.somenteConsulta}
-            >
-              <Grid cols={12}>
-                <Alert
-                  alerta={{
-                    tipo: 'warning',
-                    id: 'AlertaPrincipal',
-                    mensagem:
-                      'Você tem apenas permissão de consulta nesta tela.',
-                    estiloTitulo: { fontSize: '18px' },
-                  }}
-                />
-              </Grid>
-            </Row>
-          </div>
+          <Mensagens somenteConsulta={menuRetraido.somenteConsulta} />
         </main>
       </div>
       <Switch>
